feat(intersect): Add Methods to Intersect Solids
This commit adds methods for intersecting solid breps with one another to create matching surfaces for energy models. It adds a new bounding box check that greatly reduces the run time from the honeybee legacy version.  The single-core calculation runs in under half the time of honeybee-legacy and the parallel version runs in 1/10th of the time on my 12-core machine.