feat(dotnet): Merge ladybug_dotnet dependency into this package
When we originally set up the ladybug_dotnet repo, the idea was to pull out any functions that could be shared by a Grasshopper and a Dynamo Plugin into one package. I now know that this was a faulty assumption because:

1) It's clear that the sum-total of all code with the potential to be shared between these two plugins is less than 100 lines of code.
2) It doesn't seem possible that even these 100 lines could be shared given that Grasshopper and Dynamo are using two different versions of IronPython and most of these 100 lines rely on updated download encryption libraries that I am sure Dynamo lacks.

So, for the sake of cleaner management of dependencies, I am merging these 100 lines of ladybug_dotnet code into ladybug _rhino and I will archive the ladybug_dotnet package.