feat(pythonpath): Ensure IronPython path is set for Rhino 7 if instaled
This commit updates the set-python-search command to work for both Rhino 6 and Rhino 7.  The command is smart enough to sense what is installed and just set the path for that.

Right now, the process of setting the path for Rhino 7 is the same as it is for Rhino 6, including the copying of files to the scripts folder on Mac.  If McNeel adds a better way to do it, we can switch to that.