feat(grasshopper): Add methods for wrapping data in Python objects
Similar to what legacy did with the cumulative sky matrix, this method can wrap large amounts of data so that it doesn't slow down the Grasshopper UI. However, this method is intended to be generic for many use cases and the objects can also be de-serialized such that users can get access to the detailed data if they want it.

Using this on the sun_is_visible output for some direct sunlight studies resulted in the run time dropping to 30% for an annual study.