feat(diff): Put all functions that evaluate differences in one module
This organization seems a lot cleaner to me even though I know it mixes some of the functions used to check versions before export with the functions that check versions for syncing Grasshopper files.

The good news is that, in the process, I have definitely removed some redundant code and made it so that we have one good method for checking something rather than 2 slightly different copies of it.