feat(versioning): Have sync component update inputs + outputs
I did several tests and it looks like there shouldn't be any issues setting the type hint, item/list access, or the name change of an input or an output.

Now, the only thing that needs to be manually replaced is the case where the number of inputs or outputs have changed.