fix(geometry): Add extra check for duplicate vertices
It seems that it's possible to have a Brep where the object doesn't have duplicate vertices but the loops of the Brep's faces have them.  So this commit adds an extra check to remove such duplicate vertices.

I am also updating the polyface translators to used the more robust meshing translation whenever the input Brep has a curved face since this should help preserve solidity in more cases of using the default components without the "Planarize Brep" component.

Lastly, I am changing some of the methods so that I can remove the RhinoCommon import statement in the "Planarize Brep" component.