fix(togeometry): Catch the case of zero-area mesh faces
While Rhino has robust checks that prevent most Breps from having 0 area, the same can't really be said for meshes. So I'm adding a zero-area for the mesh case just to be sure. It has a negligible effect on the run-time since we're already computing the area to ensure the vertices are counter-clockwise.