fix(intersect): Implement better solid intersection method
This commit changes the solid intersection method to use the same underlying method as Mingbo used here in IronBug:

https://github.com/MingboPeng/Ironbug/blob/master/src/Ironbug.LBHB_Legacy/Honeybee/Honeybee_IntersectMassII.cs

It seems there was a much more elegant way to do this in RhinoCommon and I just needed to be made aware of this ShrinkFaces method.

Now, the new intersection component should be as reliable as IronBug and run in roughly half the time.