fix(togeometry): Preserve Plane orientation in translation if possible
This commit ensures that all of the information about the orientation of a planar Brep surface in Rhino is preserved when translating to ladybug_geometry.

This will be useful in workflows where one wants to ensure point grids are generated with a certain orientation or the up_vector of apertures conforms to what is seen in Rhino.

This change does not affect geometry translation from Meshes or curved Breps.