feat(grasshopper): Add methods for popups and recommended cpu count
This will make it easier to communicate certain messages to users and to set up parallel operations in a way that makes sense for the system running the calculation.

I'm also exposing a cpu_count option on the parallel solid intersection, which I plan to use in the Grasshopper plugin.