fix: remove cached version of the active doc
We currently get the active doc once when we import the libraries. This causes issues for baking the visualization sets from the panel. Here is how you can recreate the issue.

1. Open Rhino and Pollination Panel.
2. Bake the visualization set for a study.
3. Open a new file.
4. Try to bake the visualization set in the new file.

You will get an error: Could not find Layer with ID 00000000-0000-0000-0000-000000000000

This is happening because when the code tries to look up the layers and create them the `doc` variable is still referring to the previous Rhino document which had the layer and is now closed. Since the older document already had this layer, Rhino doesn't create the new layer and then when it tries to bake the geometry to that layer it sees that the layer doesn't exist and throws an error.

This commit fixes this issue by accessing the active doc from inside every function.

It also updates the call for `FindByFullPath` to use the latest version of RhinoCommon and uses `UnserIntIndex` which is what `rhinoscriptsyntax` library is using.