fix(text): Finally fix the scaling issue when baking text
It seems that the whole reason why I needed to add this mysterious factor in all of the text baking operations is because Rhino has this really annoying factor that scales all text in Model space, which was really hard to figure out. Now, I just disable it whenever we bake text and this always bakes the text correctly.